add scale path and associated parsers for dryrun and choice of archive

+++ s1ifr/sentinel1_pieuvre.py
@@ -41,7 +41,12 @@ EXTENSION_SAFE = ".SAFE"
 
 
 def finalize_archiving(
-    archive_dir, unzipped_safe, final_place, ziptype="", archive_name="mpc"
+    archive_dir,
+    unzipped_safe,
+    final_place,
+    ziptype="",
+    archive_name="mpc",
+    dryrun=False,
 ):
     """
     mv the safe to archive directory then chmod
@@ -51,6 +56,7 @@ def finalize_archiving(
         final_place (str): the place the safe should be in archive if no corruption
         ziptype (str): .tar or .zip or ''
         archive_name (str):
+        dryrun (bool): True -> no move of the product
     """
     #     unziped_safe = full_path_safe.strip('.tar')
     # check that the SAFE uncompressed is not corrupted
@@ -59,21 +65,27 @@ def finalize_archiving(
     logpath = os.path.join(
         "/home1/scratch", user_run, "sentinel1_quality_check_after_unzip.txt"
     )
-    is_ok_safe = safe_checker(unzipped_safe, logpath=logpath, security_time=0)
+    if dryrun is False:
+        is_ok_safe = safe_checker(
+            unzipped_safe, logpath=logpath, security_time=0
+        )
+    else:
+        is_ok_safe = True
 
     if is_ok_safe:
         cmd = "/bin/mv -f " + unzipped_safe + " " + archive_dir
         logging.debug("command to execute %s", cmd)
-        status = subprocess.check_output(cmd, shell=True)
-        if status != 0:
-            st = os.system("chmod 775 -R " + final_place)
-            if st != 0:
-                logging.error("chmod operation on %s failed", final_place)
-                doom_flag = FAILED
+        if dryrun is False:
+            status = subprocess.check_output(cmd, shell=True)
+            if status != 0:
+                st = os.system("chmod 775 -R " + final_place)
+                if st != 0:
+                    logging.error("chmod operation on %s failed", final_place)
+                    doom_flag = FAILED
+                else:
+                    logging.debug("chmod done")
             else:
-                logging.debug("chmod done")
-        else:
-            doom_flag = FAILED
+                doom_flag = FAILED
     else:
         logging.error(
             "%s is corrupted, so we delete it to let it be re download",
@@ -84,7 +96,7 @@ def finalize_archiving(
     original_full_path = unzipped_safe + ziptype
     logging.debug("original_full_path = %s", original_full_path)
     # remove the tar or zip file that is now useless since uncompress has been done or move to quarantine
-    if os.path.exists(original_full_path) is True:
+    if os.path.exists(original_full_path) is True and dryrun is False:
         logging.info("%s has been deleted ", original_full_path)
         if os.path.isdir(original_full_path):
             shutil.rmtree(original_full_path)
@@ -98,6 +110,7 @@ def sort_one_safe(
     log_file_handler=None,
     other_archive="datarmor_mpc",
     security_second=600,
+    dryrun=False,
 ):
     """
     :input:
@@ -105,6 +118,7 @@ def sort_one_safe(
         log_file_handler (int): [optional]
         other_archive (str) : name of the archive where to store the file [optional]
         security_second (int): number minimum of seconds since last modification [optional]
+        dryrun (bool): do not uncompress nor move the product
     :output:
         doom_flag (str): tell what appended to the SAFE treated
         cpt_dupli (int): counter of duplicate
@@ -131,7 +145,9 @@ def sort_one_safe(
         if safe_basename[0:2] == "S1" and safe_basename[-5:] != EXTENSION_SAFE:
             safe_basename = safe_basename + EXTENSION_SAFE
         logging.debug("basename : %s", safe_basename)
-        archive_dir = which_archive_dir(safe_basename)
+        archive_dir = which_archive_dir(
+            safe_basename, archive_name=other_archive
+        )
         final_place = os.path.join(archive_dir, safe_basename)
         logging.debug("final path should be %s", final_place)
         flag_continue, _, _ = product_is_present_at_ifremer(
@@ -204,14 +220,20 @@ def sort_one_safe(
                     os.chdir(spool_dir)
                     cmd = "unzip -o " + full_path_safe
                     logging.debug("command: %s", cmd)
-                    try:
-                        st = subprocess.check_output(cmd, shell=True,stderr=subprocess.STDOUT, text=True)
-                    except subprocess.CalledProcessError as e:
-                        st = e.returncode
-                        logging.error(f"Error with cmd : {e}")
-                        logging.error(f"status returned : {e.returncode}")
-                        logging.error(f"cmd output : {e.output}")
-                    logging.debug("status unzip : %s", st)
+                    if dryrun is False:
+                        try:
+                            st = subprocess.check_output(
+                                cmd,
+                                shell=True,
+                                stderr=subprocess.STDOUT,
+                                text=True,
+                            )
+                        except subprocess.CalledProcessError as e:
+                            st = e.returncode
+                            logging.error(f"Error with cmd : {e}")
+                            logging.error(f"status returned : {e.returncode}")
+                            logging.error(f"cmd output : {e.output}")
+                        logging.debug("status unzip : %s", st)
                     unziped_safe = full_path_safe.strip(".zip")
                     unziped_safe = unziped_safe.replace(
                         os.path.dirname(full_path_safe), spool_dir
@@ -224,13 +246,19 @@ def sort_one_safe(
                         unziped_safe,
                         os.path.exists(unziped_safe),
                     )
-                    if os.path.exists(unziped_safe):
+                    testexistenceuncompressedsafe = os.path.exists(
+                        unziped_safe
+                    )
+                    if dryrun is True:
+                        testexistenceuncompressedsafe = True
+                    if testexistenceuncompressedsafe:
                         doom_flag = finalize_archiving(
                             archive_dir,
                             unziped_safe,
                             final_place,
                             ".zip",
                             archive_name=other_archive,
+                            dryrun=dryrun,
                         )
                     else:
                         logging.error(
@@ -290,8 +318,11 @@ def sort_one_safe(
         )
         doom_flag = UNEXISTANT
     if doom_flag == NORMAL:
-        logging.info("final path where the product is stored : %s", final_place)
+        logging.info(
+            "final path where the product is stored : %s", final_place
+        )
     logging.debug("final flag: %s", doom_flag)
+    logging.info("final path where the product is stored : %s", final_place)
     return doom_flag, cpt_dupli
 
 
@@ -310,6 +341,18 @@ def main():
         action="store",
         dest="safe",
         help="full path of a unique SAFE product to be sorted and stored",
+    )
+    parser.add_argument(
+        "--archivename",
+        choices=["datarmor_mpc", "scale"],
+        help="name of the archive to use datarmor_mpc or scale",
+    )
+    parser.add_argument(
+        "--dryrun",
+        action="store_true",
+        default=False,
+        help="do not unzip nor move product,"
+        " to be used to check final path expected",
     )
     args = parser.parse_args()
     fmt = "%(asctime)s %(levelname)s %(filename)s(%(lineno)d) %(message)s"
@@ -333,10 +376,14 @@ def main():
     if user_run != "satwave":
         logging.warning('you must run this script with user "satwave".')
     logging.info("user : %s", user_run)
-    archive_output = ["datarmor_mpc"]
+    # archive_output = ["datarmor_mpc"]
+    archive_output = [args.archivename]
     logging.info("the script will sort sentinel1 product : %s", args.safe)
     sort_one_safe(
-        args.safe, other_archive=archive_output[0], security_second=0
+        args.safe,
+        other_archive=archive_output[0],
+        security_second=0,
+        dryrun=args.dryrun,
     )
     logging.info("time to sort the data %1.1f seconds", time.time() - t0)
